Refactoring project structure
