+++ .gitignore
@@ -5,4 +5,3 @@
 /*.exe
 *.sh
 *.py
-*.ps1
