+++ src/engine/chesslogic.cpp
