+++ src/include/chesslogic.h
