+++ src/include/uci.h
@@ -1,0 +1,10 @@
+#ifndef UCI_H
+#define UCI_H
+#include "utils.h"
+#include "movegen.h"
+#include "chesslogic.h"
+#include "bot.h"
+
+void run_uci();
+
+#endif // !UCI_H
